Fix comments tests.

+++ Source/Lib/TraktApiSharp/Objects/Basic/TraktComment.cs
@@ -14,7 +14,7 @@
 
         /// <summary>Gets or sets the parent comment id, if this comment is a reply.</summary>
         [JsonProperty(PropertyName = "parent_id")]
-        public ulong? ParentId { get; set; }
+        public uint? ParentId { get; set; }
 
         /// <summary>Gets or sets the UTC datetime, when this comment was created.</summary>
         [JsonProperty(PropertyName = "created_at")]
